Revert "T0 create backend project"
