+++ .gitignore
@@ -1,3 +1,1 @@
 /.DS_Store
-**/.vs
-**/obj
