--- coffee_machine_backend/coffee_machine_backend.http
@@ -1,6 +1,0 @@
-@coffee_machine_backend_HostAddress = http://localhost:5056
-
-GET {{coffee_machine_backend_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
