--- coffee_machine_backend/coffee_machine_backend.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35219.272
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "coffee_machine_backend", "coffee_machine_backend.csproj", "{F7703255-467B-4BFC-AEDB-74145AA2648B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F7703255-467B-4BFC-AEDB-74145AA2648B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F7703255-467B-4BFC-AEDB-74145AA2648B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F7703255-467B-4BFC-AEDB-74145AA2648B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F7703255-467B-4BFC-AEDB-74145AA2648B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C5AB6C29-D5BB-4410-9236-7A450E87BB53}
-	EndGlobalSection
-EndGlobal
